Move utility classes to util package
