+++ api/src/main/java/com/inrupt/client/util/URIBuilder.java
@@ -18,7 +18,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.inrupt.client;
+package com.inrupt.client.util;
 
 import java.net.URI;
 import java.net.URISyntaxException;
