+++ jena/src/main/java/com/inrupt/client/jena/JenaBodyPublishers.java
@@ -21,7 +21,7 @@
 package com.inrupt.client.jena;
 
 import com.inrupt.client.Request;
-import com.inrupt.client.spi.IOUtils;
+import com.inrupt.client.util.IOUtils;
 
 import org.apache.jena.graph.Graph;
 import org.apache.jena.query.Dataset;
