+++ vc/src/main/java/com/inrupt/client/vc/VerifiableCredentialBodyPublishers.java
@@ -23,9 +23,9 @@ package com.inrupt.client.vc;
 import com.inrupt.client.Request;
 import com.inrupt.client.VerifiableCredential;
 import com.inrupt.client.VerifiablePresentation;
-import com.inrupt.client.spi.IOUtils;
 import com.inrupt.client.spi.JsonService;
 import com.inrupt.client.spi.ServiceProvider;
+import com.inrupt.client.util.IOUtils;
 
 import java.io.IOException;
